+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.11.30
+
+* add `SparseMatrix.min` and `SparseMatrix.max`
+* `SparseMatrix.getitem` will raise IndexError on invalid coordinates